Fix toggleProvidingOrgFields
We were using an arrow function in our code to show or hide parts of the
Budget providing organisation UI. Whilst this will work in modern
browsers in development, in production the uglifier will complain as we
are expecting modern ES6 code.

Conceptually, the code is locating the checked radio from the array of
radio button elements, a non-modern equivalent is to loop over the
buttons and find the checked one.