Integrate activities upload in the report layout
This is attempt #1 at making the activity upload pages (starting the upload, and the post-upload feedback) integral parts of the report, by adding the report headings, meta information, and tab structure.

This way users can hopefully navigate through the upload process and after it without leaving the report context.

The list of _newly_ created and updated activities is still ephemeral, living only in the context of the post-upload feedback page.

However, the list of “activities related to this report” is going to include any reportable activities, in the way we currently define that.