Set HistoricalEvent#trackable_type explicitly
In order for `HistoricalEvent#trackable_type` to be set to
`'Actual'` after:

```
HistoricalEvent.create(
  trackable: instance of Actual
)
```

we need to explicitly set the `trackable_type` of the STI
class. If not, the `HistoricalEvent#trackabletype` will be
the base class of the STI, i.e. `'Transaction'`.

The Rails docs cover this complexity in combining
polymorphic associations and Single-Table Inheritance:

https://railsdoc.github.io/classes/ActiveRecord/Associations/ClassMethods.html#module-ActiveRecord::Associations::ClassMethods-label-Polymorphic+Associations

However, the suggested technique of overriding
`HistoricalEvent#trackable_type=(class_name)` didn't work,
so here we use a 'before create' hook to set the attribute.

TODO: add a unit test