Introduce Export::AllActivityTotals
This class has been extracted out of the SpendingBreakdown report.

It mainly encapsulated the complicated query used to get all totals for
a given activity.

I wanted to be able to share the query and document the behaviour and
this is a good way to do both.

It will return all actual totals (actual spend, refunds and adjustments)
for a given activity:

- for all time when no report is passed
- up to and including the passed in report

This is how actuals in approved (historic) reports are a snapshot of the
data at the time it was reported.

The spec is quite complicated with a lot of fixtures, but I think the
underlying concept is tricky so that is only natural.

I like being able to demonstrate how the data structure works here as
this is the foundation on which all the new export work is built and is
itself the result of seeing how this query returns the values.